Fix lynx config problems
What I did not comprehend at first is that the lynx configuration file
is not the same as the file where lynx stored the local user options.

Both files have a completely different syntax and store different kinds
of settings.

Generally, all options shown as "will not be saved" ("(!)") in the
options dialog need to be set in the configuration file instead.

One problem is that lynx always find the personalized options file
~/.lynxrc, but it only reads the global configuration file /etc/lynx.cfg
by default.

In order to do configure things which require changes to the
configuration rather than the options file, lynx needs either to be
called with -cfg option or an environment variable has to be set up.

I used the former method, writing a wrapper script with the same name
which is put ahead of the binary within $PATH.

The script finds out the name of the binary and calls it with -cfg.

So everything seens to work now.

Another problem is that no-one ever uses the ~/.lynx directory directly.
Therefore dropping that symlink and replacing by direct paths to the
former symlink's target.

(cherry picked from commit f826d0dbdfdbda7a3af65c4171e6ce6e26e20be5)
(cherry picked from commit 13d4ee8176475338c9c7640977c2b84ae235f21e)